third_party/WebKit/LayoutTests/http/tests/priority-hints/link-stylesheet-initial-load.html

<script src=resources/common.js></script>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>

<link id=link1 importance=low rel=stylesheet href=../resources/dummy.css>
<link id=link2 importance=high rel=stylesheet href=../resources/dummy.css?1>
<link id=link3 importance=auto rel=stylesheet href=../resources/dummy.css?2>
<link id=link4 importance=xyz rel=stylesheet href=../resources/dummy.css?3>
<link id=link5 rel=stylesheet href=../resources/dummy.css?4>

<script>
  async_test(t => {
    addEventListener('DOMContentLoaded', t.step_func(() => {
      const msg = 'all stylesheets must be fetched by the preload scanner';
      assert_true(internals.isPreloaded(link1.href), msg);
      assert_true(internals.isPreloaded(link2.href), msg);
      assert_true(internals.isPreloaded(link3.href), msg);
      assert_true(internals.isPreloaded(link4.href), msg);
      assert_true(internals.isPreloaded(link5.href), msg);

      assert_equals(getPriority(link1.href), kVeryHigh, 'low importance on <link rel=stylesheet> must have no effect on resource load priority');
      assert_equals(getPriority(link2.href), kVeryHigh, 'high importance on <link rel=stylesheet> must have no effect on resource load priority');
      assert_equals(getPriority(link3.href), kVeryHigh, 'auto importance on <link rel=stylesheet> must have no effect on resource load priority');
      assert_equals(getPriority(link4.href), kVeryHigh, 'invalid importance on <link rel=stylesheet> must have no effect on resource load priority');
      assert_equals(getPriority(link5.href), kVeryHigh, 'missing importance on <link rel=stylesheet> must have no effect on resource load priority');

      t.done();
    }));
  }, 'importance attribute on <link rel=stylesheet>s fetched by the preload scanner does not effect the resource load priority');
</script>
